Core Tests TX Validation (#337)
* Fix gen_tx_unlock_time, gen_tx_big_version

* Fix gen_tx_input_is_not_txin_to_key

* Change convention for block naming in working tests

* Fix gen_tx_has_inputs_no_outputs, invalid_input_amount

* Fix gen_tx_input_wo_key_offsets

* Add test note about gen_tx_mixed_key_offset_not_exist

* Uncomment tests

* Unbreak broken gx_ten tests

* Fix gen_tx_key_image_not_derive_from_tx_key

* Fix gen_tx_key_image_is_invalid

* Fix gen_tx_txout_to_key_has_invalid_key

* Add note for gen_tx_output_with_zero_amount

* Fix gen_tx_signatures_are_invalid

+++ tests/core_tests/chaingen.h
@@ -430,7 +430,7 @@ class TxBuilder {
 
   /// required fields
   const std::vector<test_event_entry>& m_events;
-  const uint8_t m_hf_version;
+  const size_t m_hf_version;
   cryptonote::transaction& m_tx;
   const cryptonote::block& m_head;
   const cryptonote::account_base& m_from;
@@ -951,7 +951,7 @@ cryptonote::transaction make_deregistration_tx(const std::vector<test_event_entr
 // NOTE(loki): These macros assume hardfork version 7 and are from the old Monero testing code
 #define MAKE_TX_MIX(VEC_EVENTS, TX_NAME, FROM, TO, AMOUNT, NMIX, HEAD)                       \
   cryptonote::transaction TX_NAME;                                                           \
-  TxBuilder(VEC_EVENTS, TX_NAME, HEAD, FROM, TO, AMOUNT, cryptonote::network_version_7).build();                            \
+  TxBuilder(VEC_EVENTS, TX_NAME, HEAD, FROM, TO, AMOUNT, cryptonote::network_version_7).build(); \
   VEC_EVENTS.push_back(TX_NAME);
 
 #define MAKE_TX(VEC_EVENTS, TX_NAME, FROM, TO, AMOUNT, HEAD) MAKE_TX_MIX(VEC_EVENTS, TX_NAME, FROM, TO, AMOUNT, 9, HEAD)
